Reimplement VanillaState as int, and update tests and interfaces.

+++ src/main/java/com/ignoretheextraclub/siteswapfactory/siteswap/State.java
@@ -31,7 +31,7 @@ public interface State extends Comparable
      *
      * @return excitedness
      */
-    int excitedness();
+    long excitedness();
 
     /**
      * The set of states this state can move to from any throw.
